Add files via upload
I have designed a Naive Bayes Classifier and trained the model with clean tweet texts data and tested the test data and achieved an accuracy of 86%.
I have designed a BERT Classifier and trained the model for 5 epochs with 95% of training accuracy and 79% of validation accuracy. the model was tested on test data and achieved an accuracy of 79%. I am planning to improve the this model in future.